requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy>=2.0
passlib[bcrypt]
bcrypt<4
PyJWT
pydantic[email]
PyMySQL
Pillow
python-multipart
cryptography
matplotlib
apscheduler
mysqlclient
openai
requests
pdfplumber
pypdfium2
pypdf
requests